+++ src/compiler/context.c
@@ -10,7 +10,7 @@ CompilationUnit *unit_create(File *file)
 	CompilationUnit *unit = CALLOCS(CompilationUnit);
     unit->file = file;
 	unit->is_interface_file = str_has_suffix(file->name, ".c3i");
-	htable_init(&unit->local_symbols, 64 * 1024);
+	htable_init(&unit->local_symbols, 1024);
     return unit;
 }